Rename js to ts
