+++ src/scripts/ui/imagePreview.ts
