+++ src/scripts/ui/toggleSwitch.ts
